Add charset fixes (from upstream 2.6.8)

+++ src/main/java/info/freelibrary/marc4j/converter/impl/ReverseCodeTable.java
@@ -232,7 +232,7 @@ public abstract class ReverseCodeTable {
         for (int i = 0; i < charsetsUsed.length(); i++) {
             final char toUse = charsetsUsed.charAt(i);
 
-            if (chars.containsKey(toUse)) {
+            if (chars.containsKey((int) toUse)) {
                 return (toUse);
             }
         }
